bitbake: override syntax causes incompatible changes honister
Since the changes are not backward compatible, we mark compatibility
with honister only.

Since we now we are on honister we can remove unused
recipes for acpica and kernel as neither should be located here.
A newer version of acpica is provided by poky. linux-kernel should be
moved to layers actually building galileo and minnowboard image.

Signed-off-by: Ferry Toth <ftoth@exalondelft.nl>